test that error-free input does not call onError
At first it seems like we could test for validationErrors.length == 0
in cases where input passes all the tests. But that's not quite correct,
because validationErrors is only populated as a side-effect when the
onError handler is called. It never exists if there are no errors.

Instead, we actually need to test whether the onError handler is called.
So we clear a boolean before each test, then enable it from onError.

Of course, this required some of the test cases to be split up, which
also led to slightly clearer naming.

This edge case is unique to asynchronous programming because a normal
function will either return or throw an exception. It probably would
have been cleaner to use a node-style callback that is called exactly
once and has the signature function(error, result).